+++ packages/connectivity/lib/connectivity.dart
@@ -97,17 +97,15 @@ class Connectivity {
 
   /// Request to authorize the location service. Only on iOS.
   ///
-  /// Returns a [LocationAuthorizationStatus] if the location service has already been authorized or user authorized the location on
-  /// this request.
+  /// This method will throw a [PlatformException] on Android.
   ///
-  /// if the location information needs to be accessible all the time, set `requestAlwaysLocationUsage` to true. Note that the status
-  /// returned might not be [LocationAuthorizationStatus.authorizedAlways] even if you requested it. The user might have already chosen a location authorization
-  /// to this app.
+  /// Returns a [LocationAuthorizationStatus] after user authorized or denied the location on this request.
   ///
-  /// It will show a platform standard window of requesting a location service.
+  /// if the location information needs to be accessible all the time, set `requestAlwaysLocationUsage` to true. If user has
+  /// already granted a [LocationAuthorizationStatus.authorizedWhenInUse] prior to requesting an "always" access, it will return [LocationAuthorizationStatus.denied].
   ///
-  /// If the user declined the location service, it will never show the window to request the authorization again.
-  /// The user has to go to the settings app in the phone to enable authorization.
+  /// If the location service authorization is not determined prior to making this call, a platform standard UI of requesting a location service will pop up.
+  /// This UI will only show once unless the user re-install the app to their phone which resets the location service authorization to not determined.
   ///
   /// This method is a helper to get the location authorization that is necessary for certain functionality of this plugin.
   /// It can be replaced with other permission handling code/plugin if preferred.
@@ -127,21 +125,82 @@ class Connectivity {
   /// For example,
   /// ```dart
   /// if (Platform.isIOS) {
-  ///   LocationAuthorizationStatus status = await _connectivity.requestLocationServiceAuthorizationIfUndetermined();
-  ///   if (status == LocationAuthorizationStatus.authorizedAlways || status == LocationAuthorizationStatus.authorizedWhenInUse) {
-  ///     wifiBSSID = await _connectivity.getWifiBSSID();
-  ///   } else {
-  ///     print('location service is not authorized');
+  ///   LocationAuthorizationStatus status = await _connectivity.getLocationServiceAuthorization();
+  ///   if (status == LocationAuthorizationStatus.notDetermined) {
+  ///     status = await _connectivity.requestLocationServiceAuthorization();
   ///   }
+  ///   if (status == LocationAuthorizationStatus.authorizedAlways || status == LocationAuthorizationStatus.authorizedWhenInUse) {
+  ///     wifiBSSID = await _connectivity.getWifiName();
+  ///   } else {
+  ///     print('location service is not authorized, the data might not be correct');
+  ///     wifiBSSID = await _connectivity.getWifiName();
+  ///   }
+  /// } else {
+  ///   wifiBSSID = await _connectivity.getWifiName();
   /// }
   /// ```
-  /// This method will throw on Android.
-  Future<LocationAuthorizationStatus> requestLocationServiceAuthorizationIfUndetermined({bool requestAlwaysLocationUsage = false}) async {
+  ///
+  /// Ideally, a location service authorization should only be requested if the current authorization status is not determined.
+  ///
+  /// See also [getLocationServiceAuthorization] to obtain current location service status.
+  Future<LocationAuthorizationStatus> requestLocationServiceAuthorization(
+      {bool requestAlwaysLocationUsage = false}) async {
+    //Just `assert(Platform.isIOS)` will disable us to do dart side unit testing.
+    assert(!Platform.isAndroid);
+    final String result = await methodChannel.invokeMethod<String>(
+        'requestLocationServiceAuthorization',
+        <bool>[requestAlwaysLocationUsage]);
+    return _convertLocationStatusString(result);
+  }
+
+  /// Get the current location service authorization. Only on iOS.
+  ///
+  /// This method will throw a [PlatformException] on Android.
+  ///
+  /// Returns a [LocationAuthorizationStatus].
+  /// If the returned value is [LocationAuthorizationStatus.notDetermined], a subsequent [requestLocationServiceAuthorization] call
+  /// can request the authorization.
+  /// If the returned value is not [LocationAuthorizationStatus.notDetermined], a subsequent [requestLocationServiceAuthorization]
+  /// will not initiate another request. It will instead return the "determined" status.
+  ///
+  /// This method is a helper to get the location authorization that is necessary for certain functionality of this plugin.
+  /// It can be replaced with other permission handling code/plugin if preferred.
+  ///
+  /// Starting from iOS 13, `getWifiBSSID` and `getWifiIP` will only work properly if:
+  ///
+  /// * The app uses Core Location, and has the user’s authorization to use location information.
+  /// * The app uses the NEHotspotConfiguration API to configure the current Wi-Fi network.
+  /// * The app has active VPN configurations installed.
+  ///
+  /// If the app falls into the first category, call this method before calling `getWifiBSSID` or `getWifiIP`.
+  /// For example,
+  /// ```dart
+  /// if (Platform.isIOS) {
+  ///   LocationAuthorizationStatus status = await _connectivity.getLocationServiceAuthorization();
+  ///   if (status == LocationAuthorizationStatus.authorizedAlways || status == LocationAuthorizationStatus.authorizedWhenInUse) {
+  ///     wifiBSSID = await _connectivity.getWifiName();
+  ///   } else {
+  ///     print('location service is not authorized, the data might not be correct');
+  ///     wifiBSSID = await _connectivity.getWifiName();
+  ///   }
+  /// } else {
+  ///   wifiBSSID = await _connectivity.getWifiName();
+  /// }
+  /// ```
+  ///
+  /// See also [requestLocationServiceAuthorization] for requesting a location service authorization.
+  Future<LocationAuthorizationStatus> getLocationServiceAuthorization() async {
     //Just `assert(Platform.isIOS)` will disable us to do dart side unit testing.
     assert(!Platform.isAndroid);
     final String result = await methodChannel
-        .invokeMethod<String>('requestLocationServiceAuthorizationIfUndetermined', <bool>[requestAlwaysLocationUsage]);
+        .invokeMethod<String>('getLocationServiceAuthorization');
+    return _convertLocationStatusString(result);
+  }
+
+  LocationAuthorizationStatus _convertLocationStatusString(String result) {
     switch (result) {
+      case 'notDetermined':
+        return LocationAuthorizationStatus.notDetermined;
       case 'restricted':
         return LocationAuthorizationStatus.restricted;
       case 'denied':
@@ -170,6 +229,8 @@ ConnectivityResult _parseConnectivityResult(String state) {
 
 /// The status of the location service authorization.
 enum LocationAuthorizationStatus {
+  /// The authorization of the location service is not determined.
+  notDetermined,
 
   /// This app is not authorized to use location.
   restricted,