+++ packages/connectivity/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.4.4
 
-* Add `requestLocationServiceAuthorizationIfUndetermined` to get location information on iOS.
+* Add `requestLocationServiceAuthorization` to request location authorization on iOS.
+* Add `getLocationServiceAuthorization` to get location authorization status on iOS.
+* Update README: add more information on iOS 13 updates with CNCopyCurrentNetworkInfo.
 
 ## 0.4.3+7